refactor classes/utilities in services
